chore: :truck: Move ProgramAnalyzer from models to processors
